+++ api.planx.uk/modules/send/service/exportZip.test.ts
@@ -1,6 +1,6 @@
-import { mockLowcalSession } from "../tests/mocks/saveAndReturnMocks";
+import { mockLowcalSession } from "../../../tests/mocks/saveAndReturnMocks";
 import { buildSubmissionExportZip } from "./exportZip";
-import type { LowCalSession } from "../types";
+import type { LowCalSession } from "../../../types";
 
 jest.mock("fs", () => ({
   mkdtempSync: () => "tmpdir",
