+++ api.planx.uk/modules/send/service/uniform.ts
@@ -3,9 +3,9 @@ import { NextFunction, Request, Response } from "express";
 import { Buffer } from "node:buffer";
 import FormData from "form-data";
 import fs from "fs";
-import { markSessionAsSubmitted } from "../modules/saveAndReturn/service/utils";
+import { markSessionAsSubmitted } from "../../saveAndReturn/service/utils";
 import { gql } from "graphql-request";
-import { $api } from "../client";
+import { $api } from "../../../client";
 import { buildSubmissionExportZip } from "./exportZip";
 
 interface UniformClient {
